% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{boot.wmt}
\alias{boot.wmt}
\title{Bootstrap EGA WMT-2 Data}
\format{A list with 8 objects (see \code{\link[EGAnet]{bootEGA}})}
\usage{
data(boot.wmt)
}
\description{
\code{\link[EGAnet]{bootEGA}} Results of \code{\link[EGAnet]{wmt2}}Data
}
\details{
\code{\link[EGAnet]{bootEGA}} results using the \code{"glasso"} model and 500 iterations
of the Wiener Matrizen-Test 2 (WMT-2)
}
\examples{
data("boot.wmt")

}
\keyword{datasets}
