% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.louvain.R
\name{community.louvain}
\alias{community.louvain}
\title{Apply the Louvain Community Detection Algorithm}
\usage{
community.louvain(network, signed = FALSE, resolution = 1)
}
\arguments{
\item{network}{Matrix or \code{\link{igraph}} network object}

\item{signed}{Boolean.
Whether the standard or signed algorithm should be used.
Defaults to \code{FALSE} or standard}

\item{resolution}{Numeric (length = 1).
A parameter that adjusts modularity to allow the algorithm to
prefer smaller (\code{resolution} > 1) or larger
(0 < \code{resolution} < 1) communities.
Defaults to \code{1} (standard modularity computation).
Currently, this argument is only available for \code{"standard"}.
Future versions may allow \code{"signed"} to take advantage of
this parameter}
}
\value{
Returns each level of memberships from the Louvain algorithm
}
\description{
A specific function to apply the Louvain community detection algorithm. There
are two versions of the algorithm available: standard and signed. The standard
Louvain algorithm uses the absolute values of the network. The signed Louvain algorithm
uses the signed values of the network, which is made possible by using a signed version
of modularity (see Gomez et al., 2009).
}
\examples{
# Load data
wmt <- wmt2[,7:24]

# Estimate network
network <- EBICglasso.qgraph(data = wmt)

# Compute standard Louvain
community.louvain(network, signed = FALSE)

# Compute signed Louvain
community.louvain(network, signed = TRUE)

}
\references{
\strong{Louvain algorithm} \cr
Blondel, V. D., Guillaume, J.-L., Lambiotte, R., & Lefebvre, E. (2008).
Fast unfolding of communities in large networks.
\emph{Journal of Statistical Mechanics: Theory and Experiment}, \emph{2008}(10), P10008.

\strong{Signed modularity} \cr
Gomez, S., Jensen, P., & Arenas, A. (2009).
Analysis of community structure in networks of correlated data.
\emph{Physical Review E}, \emph{80}(1), 016114.
}
\author{
Hudson Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
