% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2igraph.R
\name{convert2igraph}
\alias{convert2igraph}
\title{Convert networks to \code{\link{igraph}}}
\usage{
convert2igraph(A)
}
\arguments{
\item{A}{Matrix or data frame.
\emph{N} x \emph{N} matrix where \emph{N} is the number of nodes}
}
\value{
Returns a network in the \code{\link{igraph}} format
}
\description{
Converts networks to \code{\link{igraph}} format
}
\examples{
convert2igraph(ega.wmt$network)

}
\author{
Hudson Golino <hfg9s at virginia.edu> & Alexander P. Christensen <alexander.christensen at Vanderbilt.Edu>
}
