% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2tidygraph.R
\name{convert2tidygraph}
\alias{convert2tidygraph}
\title{Convert networks to \code{\link{tidygraph}}}
\usage{
convert2tidygraph(EGA.object)
}
\arguments{
\item{EGA.object}{A single \code{\link{EGAnet}} object containing the outputs
\code{$network} and \code{$wc}}
}
\value{
Returns a network in the \code{\link{tidygraph}} format
}
\description{
Converts networks to \code{\link{tidygraph}} format
}
\examples{
convert2tidygraph(ega.wmt)

}
\author{
Dominique Makowski, Hudson Golino <hfg9s at virginia.edu>, & Alexander P. Christensen <alexander.christensen at Vanderbilt.Edu>
}
