% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/depression.R
\docType{data}
\name{depression}
\alias{depression}
\title{Depression Data}
\format{
A 574x78 response matrix

A 574x78 response matrix
}
\usage{
data(depression)

data(depression)
}
\description{
A response matrix (n = 574) of the Beck Depression Inventory, Beck Anxiety Inventory and
the Athens Insomnia Scale.

A response matrix (n = 574) of the Beck Depression Inventory, Beck Anxiety Inventory and
the Athens Insomnia Scale.
}
\examples{
data("depression")

data("depression")

}
\keyword{datasets}
