% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualTotalCor.R
\name{dualTotalCor}
\alias{dualTotalCor}
\title{Dual Total Correlation}
\usage{
dualTotalCor(data, base = 2.718282)
}
\arguments{
\item{data}{Matrix or data frame.
Should consist only of variables to be used in the analysis}

\item{base}{Numeric (length = 1).
Base to use for entropy.
Defaults to \code{exp(1)} or \code{2.718282}}
}
\value{
Returns a list containing:

\item{Ind.Entropies}{Individual entropies for each variable}

\item{Cond.Entropies}{Conditional entropies}

\item{Joint.Entropy}{The joint entropy of the dataset}

\item{Dual.Total.Cor}{The dual total correlation of the dataset}

\item{Normalized}{Dual total correlation divided by the joint entropy}
}
\description{
Computes the dual total correlation of a dataset
}
\examples{
# Compute dual total correlation
dualTotalCor(wmt2[,7:24])

}
\references{
\strong{Formalization of dual total correlation used} \cr
Varley, T. F., Pope, M., Faskowitz, J., & Sporns, O. (2023).
Multivariate information theory uncovers synergistic subsystems of the human cerebral cortex.
\emph{Communications Biology}, \emph{6}(1), 451.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
