% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGA.methods.section.R
\name{EGA.methods.section}
\alias{EGA.methods.section}
\title{Automated Methods Section for \code{\link[EGAnet]{EGA}} Objects}
\usage{
EGA.methods.section(EGA.object)
}
\arguments{
\item{EGA.object}{An \code{\link[EGAnet]{EGA}} object}
}
\value{
Automated HTML Methods section in your default browser
}
\description{
This function accepts an \code{\link[EGAnet]{EGA}} object
and generates a Methods section for your analysis. The output is
an HTML containing the descriptions of the methods and parameters
along with a Reference section for appropriate citation.
}
\examples{
EGA.methods.section(ega.wmt)

}
