% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_to_tidygraph.R
\name{ega_to_tidygraph}
\alias{ega_to_tidygraph}
\title{Convert EGAnet objects to tidygraph}
\usage{
ega_to_tidygraph(ega)
}
\arguments{
\item{ega}{An EGA object.}
}
\description{
Converts an EGA object to the input required by \code{tidygraph::tbl_graph()}
so that it can be plotted using \code{ggraph}.
}
\examples{
\dontrun{
  library(tidygraph)
  library(ggraph)

  ega.wmt <- EGA(wmt2[,7:24], plot.EGA = FALSE)

  x <- ega_to_tidygraph(ega.wmt)


  graph <- tidygraph::tbl_graph(nodes = x$nodes, edges = x$edges)

  ggraph::ggraph(graph) +
   geom_edge_link(aes(colour = link)) +
   geom_node_point(aes(colour = dimension))
}
}
