% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGM.compare.R
\name{EGM.compare}
\alias{EGM.compare}
\title{Compare \code{\link[EGAnet]{EGM}} with Exploratory Factor Analysis}
\usage{
EGM.compare(data, rotation = "geominQ", ...)
}
\arguments{
\item{data}{Matrix or data frame.
Should consist only of variables to be used in the analysis.
Can be raw data or a correlation matrix}

\item{...}{Additional arguments to be passed on to
\code{\link[EGAnet]{auto.correlate}},
\code{\link[EGAnet]{network.estimation}},
\code{\link[EGAnet]{community.detection}},
\code{\link[EGAnet]{community.consensus}},
\code{\link[EGAnet]{community.unidimensional}},
\code{\link[EGAnet]{EGA}}, and
\code{\link[psych]{fa}}}
}
\description{
Estimates an \code{\link[EGAnet]{EGM}} based on \code{\link[EGAnet]{EGA}} and
uses the number of communities as the number of dimensions in exploratory factor analysis
(EFA) using \code{\link[psych]{fa}}
}
\examples{
# Compare EGM with EFA
EGM.compare(wmt2[,7:24])

}
\author{
Hudson F. Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
