% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyFit.R
\name{entropyFit}
\alias{entropyFit}
\title{Fit of Factor Model using Empirical Entropy}
\usage{
entropyFit(data, structure, bins = 10)
}
\arguments{
\item{data}{A dataset}

\item{structure}{A vector representing the structure (numbers or labels for each item).
Can be theoretical factors or the structure detected by \code{\link{EGA}}}

\item{bins}{An integer, indicating thhe number of bins to use.}
}
\value{
Returns the values of entropy per factor, the mean empirical entropy (i.e., the average entropy across factors),
the minimum joint entropy (i.e. joint entropy for all variables) and the adjusted entropy (ratio between joint entropy and
the minumum joint entropy).
}
\description{
Uses the \code{\link[entropy]{entropy.empirical}} function
from the \code{\link{entropy}} package to compute the entropy of a
factor model. Mutual information (\code{\link[entropy]{mi.plugin}})
is used to derive associations between factors. The final \code{entropyFit}
value is entropy + mutual information.
Lower values suggest better fit with the data
}
\examples{
ega.wmt <- EGA(data = wmt2[,7:24], model = "glasso")

entropyFit(data = wmt2[,7:24], structure = ega.wmt$wc, bins = 10)

}
\seealso{
\code{\link{EGA}} to estimate the number of dimensions of an instrument using EGA and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu> and Alexander Christensen <alexpaulchristensen@gmail.com>
}
