% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_inference.GGMnonreg.R
\name{ggm_inference.GGMnonreg}
\alias{ggm_inference.GGMnonreg}
\title{\code{ggm_inference} from \code{GGMnonreg} 1.0.0}
\usage{
ggm_inference.GGMnonreg(
  data,
  alpha = 0.1,
  boot = TRUE,
  iter = 100,
  corr = c("auto", "cor_auto", "pearson", "spearman"),
  na.data = c("pairwise", "listwise"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Matrix or data frame.
Should consist only of variables to be used in the analysis.
Raw data must be provided (i.e., no correlation matrices)}

\item{alpha}{Numeric (length = 1).
Significance level for edge inclusion.
Defaults to \code{0.10} based on Williams and Rodriguez (2022)}

\item{boot}{Boolean (length = 1).
Defaults to \code{TRUE}.
Can only be \code{FALSE} for Pearson's
(\code{corr = "pearson"}) and Spearman's
(\code{corr = "spearman"}) correlations}

\item{iter}{Numeric (length = 1).
Number of replica samples to generate from the bootstrap analysis.
Defaults to \code{100} (minimum allowed).
Default in \code{GGMnonreg} is \code{1000}}

\item{corr}{Character (length = 1).
Method to compute correlations.
Defaults to \code{"auto"}.
Available options:

\itemize{

\item \code{"auto"} --- Automatically computes appropriate correlations for
the data using Pearson's for continuous, polychoric for ordinal,
tetrachoric for binary, and polyserial/biserial for ordinal/binary with
continuous. To change the number of categories that are considered
ordinal, use \code{ordinal.categories}
(see \code{\link[EGAnet]{polychoric.matrix}} for more details)

\item \code{"cor_auto"} --- Uses \code{\link[qgraph]{cor_auto}} to
compute correlations. Arguments can be passed along to the function

\item \code{"pearson"} --- Pearson's correlation is computed for
all variables regardless of categories

\item \code{"spearman"} --- Spearman's rank-order correlation is
computed for all variables regardless of categories

}}

\item{na.data}{Character (length = 1).
How should missing data be handled?
Defaults to \code{"pairwise"}.
Available options:

\itemize{

\item \code{"pairwise"} --- Computes correlation for all available
cases between two variables

\item \code{"listwise"} --- Computes correlation for all complete
cases in the dataset

}}

\item{verbose}{Boolean (length = 1).
Whether messages and (insignificant) warnings should be output.
Defaults to \code{FALSE} (silent calls).
Set to \code{TRUE} to see all messages and warnings for every function call}

\item{...}{Not actually used but makes it easier for general functionality
in the package}
}
\value{
A partial correlation matrix
}
\description{
This minimal implementation of the \code{ggm_inference} function in
\code{GGMnonreg} is modified to operate with \code{\link{EGAnet}} with maximum efficiency
(i.e., significantly faster computation for categorical correlations). The
goal is to maintain identical functionality to the original function
but faster (for more details, see Williams et al., 2019)
}
\examples{
# Obtain data
wmt <- wmt2[,7:24]

\dontrun{
# Compute non-regularized network
nonreg <- ggm_inference.GGMnonreg(data = wmt)}

}
\references{
\strong{Seminal paper on Maximum Likelihood Estimation} \cr
Williams, D. R., Rhemtulla, M., Wysocki, A. C., & Rast, P. (2019).
On nonregularized estimation of psychological networks.
\emph{Multivariate Behavioral Research}, \emph{54}(5), 719-750.

\strong{Original implementation (GGMnonreg package)} \cr
Williams, D. R. (2021).
GGMnonreg: Non-regularized Gaussian Graphical Models in R.
\emph{Journal of Open Source Software}, \emph{6}(67), 3308.

\strong{(Non-)regularization network generalizability} \cr
Williams, D. R., & Rodriguez, J. E. (2022).
Why overfitting is not (usually) a problem in partial correlation networks.
\emph{Psychological Methods}, \emph{27}(5), 822-840.
}
\author{
Donald R. Williams; for maintanence,
Hudson Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen at gmail.com>
}
