% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/intelligenceBattery.R
\docType{data}
\name{intelligenceBattery}
\alias{intelligenceBattery}
\title{Intelligence Data}
\format{
A 1185x125 response matrix

A 1185x125 response matrix
}
\usage{
data(intelligenceBattery)

data(intelligenceBattery)
}
\description{
A response matrix (n = 1152) of the International Cognitive Ability Resource (ICAR)
intelligence battery developed by Condon and Revelle (2016).

A response matrix (n = 1152) of the International Cognitive Ability Resource (ICAR)
intelligence battery developed by Condon and Revelle (2016).
}
\examples{
data("intelligenceBattery")

data("intelligenceBattery")

}
\keyword{datasets}
