% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.ident.R
\name{item.ident}
\alias{item.ident}
\title{Item Dimension Identification}
\usage{
item.ident(bootega.obj, confirm, rep.val = 0.8)
}
\arguments{
\item{bootega.obj}{A \code{\link{bootEGA}} object}

\item{confirm}{A vector with community numbers or labels for each item}

\item{rep.val}{A replication value between 0 and 1. It's recommended
to first run \code{\link{item.confirm}} to determine appropriate cut-off
value.
Defaults to .80}
}
\value{
Returns a list containing:

\item{unstd}{The unstandardized within- and between-community
strength values for each node}

\item{std}{The standardized (proportion of strength)
within- and between-community strength values for each node}
}
\description{
Computes the within- and between-community strength of each item
for each community. Based on \code{\link{item.confirm}}, researchers can flag
items that are not replicating well to identify which communities they are
falling between.
}
\examples{
\dontrun{
ega.wmt <- EGA(data = wmt2[,7:24], model = "glasso")

boot.wmt <- bootEGA(data = wmt2[,7:24], n = 100, typicalStructure = TRUE,
plot.typicalStructure = TRUE, model = "glasso",
type = "parametric", ncores = 4, confirm = ega.wmt$wc)

item.confirm(boot.wmt, confirm = ega.wmt$wc, plot.ic = TRUE)

item.ident(boot.wmt, confirm = ega.wmt$wc, rep.val = .80)
}

}
\author{
Hudson F. Golino <hfg9s at virginia.edu> and Alexander Christensen <alexpaulchristensen@gmail.com>
}
