% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemConfirm.R
\name{itemConfirm}
\alias{itemConfirm}
\title{Plot the replicability of EGA item-wise.}
\usage{
itemConfirm(bootega.obj)
}
\arguments{
\item{bootega.obj}{A bootEGA object}
}
\description{
\code{itemConfirm} Based on the bootEGA results, this function plots the number of times an item (variable) is estimated in the same
factor/dimension as originaly estimated by EGA.
}
\examples{
\dontrun{
ega.wmt <- EGA(data = wmt2[,7:24], model = "glasso")
boot.wmt <- bootEGA(data = wmt2[,7:24], n = 100, typicalStructure = TRUE,
plot.typicalStructure = TRUE, model = "GGM",
type = "parametric", ncores = 4, confirm = ega.wmt$wc)
itemConfirm(boot.wmt)
}

\dontrun{
itemConfirm(bootEGA)
}
}
\seealso{
\code{\link{EGA}} to estimate the number of dimensions of an instrument using EGA and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
