% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemDiagnostics.R
\name{itemDiagnostics}
\alias{itemDiagnostics}
\title{Diagnostics Analysis for Low Stability Items}
\usage{
itemDiagnostics(data, minor.method = c("cosine", "residuals"), ...)
}
\arguments{
\item{data}{Matrix or data frame.
Should consist only of variables to be used in the analysis}

\item{minor.method}{Character (length = 1).
Method to detect the likelihood of minor dimensions.
Available options include:

\itemize{

\item \code{"cosine"} (default) --- uses \code{\link[EGAnet]{cosine}} similarity to determine
the similarity of the item stability patterns from \code{\link[EGAnet]{bootEGA}}.
More consistent patterns across multiple dimensions reflect an increased likelihood
of a minor dimension (particularly with respect to extra dimensions forming).
By default, values greater than or equal to \code{0.95} are selected

\item \code{"residuals"} --- subtracts the network-implied zero-order correlation matrix
from the empirical zero-order correlation matrix to determine residuals. These residuals
are submitted to an \code{\link[EGAnet]{EGA}} and network loadings (\code{\link[EGAnet]{net.loads}})
are re-computed. By default, loadings greater than or equal to \code{0.35} are selected

}}

\item{...}{Additional arguments to pass on to
\code{\link[EGAnet]{bootEGA}},
\code{\link[EGAnet]{net.loads}}, and
\code{\link[EGAnet]{UVA}}}
}
\value{
Returns a list containing:

\item{diagnostics}{A data frame containing the diagnostics of low item stabilities
(see \code{\link[EGAnet]{itemStability}})}

\item{boot}{Output from \code{\link[EGAnet]{bootEGA}}}

\item{uva}{Output from \code{\link[EGAnet]{UVA}}}

\item{minor}{A list containing suggested items to \code{keep}, \code{remove}, and
a matrix for probable minor dimensions (\code{minor.matrix})}

\item{loadings}{Output from \code{\link[EGAnet]{net.loads}}}

\item{suggested}{Variables that are suggested to be retained to increase item stability}
}
\description{
Computes the between- and within-community
\code{strength} of each variable for each community
}
\examples{
# Load data
wmt <- wmt2[,7:24]

\dontrun{
# Obtain diagnostics
diagnostics <- itemDiagnostics(wmt, ncores = 2)}

}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>, Hudson Golino <hfg9s at virginia.edu>, and Luis Eduardo Garrido <garrido.luiseduardo@gmail.com>
}
