% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knifejack.ergoInfo.R
\name{knifejack.ergoInfo}
\alias{knifejack.ergoInfo}
\title{The Knife-Jack Test for the Ergodicity Information Index}
\usage{
knifejack.ergoInfo(
  dynEGA.object = NULL,
  EII,
  use = c("edge.list", "weights"),
  ncores
)
}
\arguments{
\item{dynEGA.object}{A \code{\link[EGAnet]{dynEGA}} or a
\code{\link[EGAnet]{dynEGA.pop.ind}} object that is used to match the arguments of the EII object.}

\item{EII}{A \code{\link[EGAnet]{ergoInfo}} object, used to estimate the Empirical Ergodicity Information Index, or the estimated value of EII estimated
using the \code{\link[EGAnet]{ergoInfo}} function.}

\item{use}{Character.
A string indicating what network element will be used to compute the algorithm complexity in the \code{\link[EGAnet]{ergoInfo}} function,
the list of edges or the weights of the network.
Defaults to \code{use = "edge.list"}.
Current options are:

\itemize{

\item{\strong{\code{edge.list}}}
{Calculates the algorithm complexity using the list of edges.}

\item{\strong{\code{weights}}}
{Calculates the algorithm complexity using the weights of the network.}
}}

\item{ncores}{Numeric.
Number of cores to use in computing results.
Defaults to \code{parallel::detectCores() / 2} or half of your
computer's processing power.
Set to \code{1} to not use parallel computing.
Recommended to use maximum number of cores minus one

If you're unsure how many cores your computer has,
then use the following code: \code{parallel::detectCores()}}
}
\value{
Returns a list containing:

\item{knifejack.ergoInfo}{The values of the Ergodicity Information Index obtained in the Knife-Jack sampling}

\item{p.value.twosided}{The p-value of the Monte-Carlo test for the Ergodicity Information Index.
The null hypothesis is that the empirical Ergodicity Information index is equal to the expected value of the EII if the all individuals
had similar latent structures.}

\item{effect}{Indicates wheter the empirical EII is greater or less then the Monte-Carlo obtained EII.}

\item{plot.dist}{Histogram of the bootstrapped ergodicity information index}
}
\description{
Tests the Ergodicity Information Index obtained in the empirical sample with the EII obtained by inverse-Jackkife sampling. In traditional Jackknife
sampling, one single observation is omitted in each data sample. Here, we're using the inverse of that, meaning that we analyze one individual at a time. Therefore,
since for each individual \code{i} the individual and population structure is the same (because is a single individual structure), the ergodicity information index obtained
is what you could expect if every individual in a dataset had the same structure as \code{i}. Thereforem the distribution of EII is the null distribution obtained under the hypothesis
that the structure of individuals are uniquely similar to a population structure. Since no individual is exactly equal to one another, there's enough variability in the Knife-Jack distribution.
If all individuals have the same structure, then there would be no significant difference between the empirical EII value and the null distribution of EII values obtained by the Knife-Jack approach.
However, if individuals have different structures, then the empirical EII value would be significantly different to the null distribution of EII values obtained via Knife-Jaccking.
The p-values in the Knife-Jack test can be calculated as \code{(sum(EII>=knifejack.EII)+1)/(N+1)} and as
\code{(sum(EII<=knifejack.EII)+1)/(N+1)}, where EII is the empirical Ergodicity Information Index, knifejack.EII is the values of the Ergodicity Information Index obtained
in the Knife-Jack samples, and \code{N} is the number of unique individuals in the dataset. The two-sided p-value is computed as two times the lowest p-value. In the Knife-Jack Test for the Ergodicity Information Index,
the null hypothesis is that the empirical value of EII is equal to the values of EII obtained in multiple individuals with the same structure as the population structure estimated
via \code{\link[EGAnet]{dynEGA}}.
Small values of p indicate that is very unlikely to obtain an EII as large as the one obtained in the empirical sample if the null hypothesis is true (i.e. all individuals have the same structure as the population structure), thus there is convincing evidence that the empirical Ergodicity Information Index is
different than it could be expected if all individuals had a similar latent structure.
}
\examples{

\dontrun{
\donttest{
dyn1 <- dynEGA.ind.pop(data = sim.dynEGA[,-c(22)], n.embed = 5, tau = 1,
                      delta = 1, id = 21, use.derivatives = 1,
                    model = "glasso", ncores = 2, corr = "pearson")

eii1 <- ergoInfo(dynEGA.object = dyn1)$EII

testing.ergoinfo <- knifejack.ergoInfo(
  dynEGA.object = dyn1,
  EII = eii1,
  use = "edge.list",
  ncores = 2)
}}

}
\author{
Hudson Golino <hfg9s at virginia.edu> & Alexander P. Christensen <alexander.christensen at Vanderbilt.Edu>
}
