% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.section.R
\name{methods.section}
\alias{methods.section}
\title{Automated Methods Section for \code{\link{EGAnet}} Objects}
\usage{
methods.section(
  ...,
  stats = c("net.loads", "net.scores", "dimStability", "itemStability", "entropyFit",
    "tefi", "vn.entropy", "ergoInfo", "mctest.ergoInfo")
)
}
\arguments{
\item{...}{\code{\link{EGAnet}} objects.
Available methods:

\itemize{

\item{\code{\link[EGAnet]{EGA}}}
{Exploratory graph analysis}

\item{\code{\link[EGAnet]{bootEGA}}}
{Bootstrap exploratory graph analysis}

}}

\item{stats}{Methods section for statistics in \code{\link{EGAnet}}.
Multiple statistics can be input.
Available statistics:

\itemize{

\item{\code{\link[EGAnet]{net.loads}}}
{Network loadings. Requires \link[EGAnet]{EGA} object to be input}

\item{\code{\link[EGAnet]{net.scores}}}
{Network scores. Requires \link[EGAnet]{EGA} object to be input}

\item{\code{\link[EGAnet]{dimStability}}}
{Structural consistency. Requires \link[EGAnet]{bootEGA} object to be input}

\item{\code{\link[EGAnet]{itemStability}}}
{Item stability. Requires \link[EGAnet]{bootEGA} object to be input}

}}
}
\value{
Automated HTML Methods section in your default browser
}
\description{
This function accepts \code{\link[EGAnet]{EGA}} objects
and generates a Methods section for your analysis. The output is
an HTML containing the descriptions of the methods and parameters
along with a Reference section for appropriate citation.
}
\examples{
methods.section(ega.wmt)

}
