% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlnEGA.R
\name{mlnEGA}
\alias{mlnEGA}
\title{Multilayer Behavioral and Brain Networks}
\usage{
mlnEGA(
  behav.data,
  behav.time = FALSE,
  brain.data,
  brain.time = FALSE,
  alpha = 0.05,
  ncores,
  EGA.scores = TRUE,
  EGA.args = list(),
  plot.mln = TRUE,
  plot.args = list()
)
}
\arguments{
\item{behav.data}{Matrix or data frame.
Behavioral data where rows are participants and
columns are variables}

\item{behav.time}{Boolean.
Is the behavioral data time series?
Defaults to \code{FALSE}}

\item{brain.data}{Array or list.
Entries should be correlation matrices that correspond
to the participants in \code{behav.data}}

\item{brain.time}{Boolean.
Is the brain data time series?
Defaults to \code{FALSE}}

\item{alpha}{Numeric.
Significance level for interlayer links.
Defaults to \code{.05}}

\item{ncores}{Numeric.
Number of cores to use in computing results.
Defaults to \code{parallel::detectCores() / 2} or half of your
computer's processing power.
Set to \code{1} to not use parallel computing}

\item{EGA.scores}{Boolean.
Should \code{\link[EGAnet]{EGA}} be used to determine
dimensions and compute network scores (\code{\link[EGAnet]{net.scores}})?
Defaults to \code{TRUE}.
Uses \code{\link[EGAnet]{dynEGA}} for time series data.
Set to \code{FALSE} if \code{behav.data} is already scored}

\item{EGA.args}{List.
A list of additional arguments for computing EGA.
See \code{\link[EGAnet]{EGA}} for cross-sectional data and
\code{\link[EGAnet]{dynEGA}} for time series data}

\item{plot.mln}{Boolean.
If \code{TRUE}, returns a plot of the multilayer network.
Defaults to \code{TRUE}}

\item{plot.args}{List.
A list of additional arguments for the network plot:

\itemize{

\item{\strong{\code{vsize}}}
{Size of the nodes. Defaults to 6.}

\item{\strong{\code{label.size}}}
{Size of the labels. Defaults to 5.}

\item{\strong{\code{alpha}}}
{The level of transparency of the nodes, which might be a single value or a vector of values. Defaults to 0.4.}

\item{\strong{\code{edge.alpha}}}
{The level of transparency of the edges, which might be a single value or a vector of values. Defaults to 0.7.}

}}
}
\value{
Returns a list containing:

\item{behavioral.scores}{Either scores input as \code{data} or network scores estimated from
\code{\link[EGAnet]{EGA}}}

\item{brain.networks}{A list containing:

\itemize{

\item{\code{individuals}}
{A list containing the correlation matrix of each individual participant}

\item{\code{group}}
{A matrix of the group correlation matrix (mean of each ROI--ROI connection across participants)}

\item{\code{group.threshold}}
{The threshold group brain matrix. The threshold is the critical correlation value
based on the \code{alpha} that was input}

}

}

\item{behavioral.network}{A network estimation using the \code{behavioral.scores} and \code{\link[qgraph]{EBICglasso}}}

\item{interlayer.links}{Correlations between the \code{behavioral.scores} and node \code{\link[NetworkToolbox]{strength}}
of each node in the \code{brain.networks$individuals} network}

\item{multilayer.network}{The adjacency matrix of the multilayer network that combines the \code{behavioral.network},
\code{interlayer.links}, and \code{brain.networks$group.threshold} networks}
}
\description{
EXPERIMENTAL -- USE WITH CAUTION.
Computes a multilayered network between behavioral and brain networks
following Brooks et al. (2020; see references)
}
\references{
Brooks, D., Hulst, H. E., de Bruin, L., Glas, G., Geurts, J. J. G., & Douw, L. (2020).
The multilayer network approach in the study of personality neuroscience.
\emph{Brain Sciences}, \emph{10}, 915.
doi:\href{https://doi.org/10.3390/brainsci10120915}{10.3390/brainsci10120915}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
