% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptives.R
\name{network.descriptives}
\alias{network.descriptives}
\title{Descriptive Statistics for Networks}
\usage{
network.descriptives(network)
}
\arguments{
\item{network}{Matrix, data frame,
\code{\link[qgraph]{qgraph}}, or \code{\link[EGAnet]{EGA}} object}
}
\value{

}
\description{
Computes descriptive statistics for network models
}
\examples{
# Load data
wmt <- wmt2[,7:24]

\donttest{# EGA example
## plot.type = "qqraph" used for CRAN checks
## plot.type = "GGally" is the default
ega.wmt <- EGA(data = wmt, plot.type = "qgraph")
}

# Compute descriptives
network.descriptives(ega.wmt)

}
\references{
# swn.HG \cr
Humphries, M. D., & Gurney, K. (2008).
Network 'small-world-ness': A quantitative method for determining canonical network equivalence.
\emph{PLoS one}, \emph{3}, e0002051

# swn.TJHBL \cr
Telesford, Q. K., Joyce, K. E., Hayasaka, S., Burdette, J. H., & Laurienti, P. J. (2011).
The ubiquity of small-world networks.
\emph{Brain Connectivity}, \emph{1}(5), 367-375

# scale-free_R-sq \cr
Langfelder, P., & Horvath, S. (2008).
WGCNA: an R package for weighted correlation network analysis.
\emph{BMC Bioinformatics}, \emph{9}, 559
}
\author{
Hudson Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
