% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.entropyWalk.R
\name{network.entropyWalk}
\alias{network.entropyWalk}
\title{Entropy Walk Penalized GLASSO}
\usage{
network.entropyWalk(
  data,
  n = NULL,
  corr = c("auto", "cor_auto", "cosine", "pearson", "spearman"),
  na.data = c("pairwise", "listwise"),
  gamma = 0,
  nlambda = 100,
  lambda.min.ratio = 0.1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Matrix or data frame.
Should consist only of variables to be used in the analysis}

\item{n}{Numeric (length = 1).
Sample size \strong{must} be provided if \code{data} provided is a correlation matrix}

\item{corr}{Character (length = 1).
Method to compute correlations.
Defaults to \code{"auto"}.
Available options:

\itemize{

\item \code{"auto"} --- Automatically computes appropriate correlations for
the data using Pearson's for continuous, polychoric for ordinal,
tetrachoric for binary, and polyserial/biserial for ordinal/binary with
continuous. To change the number of categories that are considered
ordinal, use \code{ordinal.categories}
(see \code{\link[EGAnet]{polychoric.matrix}} for more details)

\item \code{"cor_auto"} --- Uses \code{\link[qgraph]{cor_auto}} to compute correlations.
Arguments can be passed along to the function

\item \code{"cosine"} --- Uses \code{\link[EGAnet]{cosine}} to compute cosine similarity

\item \code{"pearson"} --- Pearson's correlation is computed for all
variables regardless of categories

\item \code{"spearman"} --- Spearman's rank-order correlation is computed
for all variables regardless of categories

}

For other similarity measures, compute them first and input them
into \code{data} with the sample size (\code{n})}

\item{na.data}{Character (length = 1).
How should missing data be handled?
Defaults to \code{"pairwise"}.
Available options:

\itemize{

\item \code{"pairwise"} --- Computes correlation for all available cases between
two variables

\item \code{"listwise"} --- Computes correlation for all complete cases in the dataset

}}

\item{gamma}{Numeric (length = 1)
EBIC tuning parameter.
Defaults to \code{0} and is generally a good choice.
Setting to \code{0} will cause regular BIC to be used}

\item{nlambda}{Numeric (length = 1).
Number of lambda values to test.
Defaults to \code{100}}

\item{lambda.min.ratio}{Numeric (length = 1).
Ratio of lowest lambda value compared to maximal lambda.
Defaults to \code{0.1}.
\strong{NOTE} \code{qgraph} sets the default to \code{0.01}}

\item{verbose}{Boolean (length = 1).
Whether messages and (insignificant) warnings should be output.
Defaults to \code{FALSE} (silent calls).
Set to \code{TRUE} to see all messages and warnings for every function call}

\item{...}{Additional arguments to be passed on to \code{\link[EGAnet]{auto.correlate}}}
}
\description{
The graphical least absolute shrinkage and selection operator with
adaptive penalization based on the random walk transition matrix of the empirical
partial correlation matrix
}
\examples{
# Obtain data
wmt <- wmt2[,7:24]

# Obtain network
RW_network <- network.entropyWalk(data = wmt)

}
\author{
Alexander P. Christensen <alexpaulchristensen at gmail.com> and
Hudson Golino <hfg9s at virginia.edu>
}
