% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.importance.R
\name{network.importance}
\alias{network.importance}
\title{Network (Relative) Importance}
\usage{
network.importance(data, network, ordinal.categories = 7, iter = 1)
}
\arguments{
\item{data}{Matrix or data frame.
Should consist only of variables to be used in the analysis.
Must be raw data (no correlation matrix)}

\item{network}{Matrix or data frame.
A partial correlation network associated with \code{data}}

\item{ordinal.categories}{Numeric (length = 1).
\emph{Up to} the number of categories \emph{before} a variable is considered continuous.
Defaults to \code{7} categories before \code{8} is considered continuous}

\item{iter}{Numeric (length = 1).
Number of shuffles to perform for each node.
Defaults to \code{1}}
}
\value{
A list containing:

\itemize{

\item results --- average of the average change in predictive power over all iterations

\item full --- a list of the average change in predictive power for each iteration

}
}
\description{
Computes node-wise relative importance by permutating each
variable in the dataset to determine the average effect on prediction power
}
\examples{
# Load data
wmt <- wmt2[,7:24]

# EBICglasso (default for EGA functions)
glasso_network <- network.estimation(data = wmt, model = "glasso")

# Check importance
network.importance(data = wmt, network = glasso_network)

}
\seealso{
\code{\link[EGAnet]{network.predictabilty}} for the underlying function
and \code{\link[EGAnet]{network.generalizability}} for additional generalizability functionality
}
\author{
Hudson Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
