% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oinformation.R
\name{Oinformation}
\alias{Oinformation}
\title{O-information}
\usage{
Oinformation(data, base = 2.718282)
}
\arguments{
\item{data}{Matrix or data frame.
Should consist only of variables to be used in the analysis}

\item{base}{Numeric (length = 1).
Base to use for entropy.
Defaults to \code{exp(1)} or \code{2.718282}}
}
\value{
Returns a list containing:

\item{Total.Cor}{The total correlation of the dataset}

\item{Dual.Total.Cor}{The dual total correlation of the dataset}

\item{O.Information}{The O-information of the dataset}

\item{Normalized.Total.Cor}{Normalized total correlation}

\item{Normalized.Dual.Total.Cor}{Normalized dual total correlation}

\item{Normalized.O.Information}{Normalized O-information}
}
\description{
Computes the O-information of a dataset
}
\examples{
# Compute O-information
Oinformation(wmt2[,7:24])

}
\references{
\strong{Formalization of O-information} \cr
Crutchfield, J. P. (1994). The calculi of emergence: Computation, dynamics and induction.
\emph{Physica D: Nonlinear Phenomena}, \emph{75}(1-3), 11-54.

\strong{Applied implementation} \cr
Marinazzo, D., Van Roozendaal, J., Rosas, F. E., Stella, M., Comolatti, R., Colenbier, N., Stramaglia, S., & Rosseel, Y. (2024).
An information-theoretic approach to build hypergraphs in psychometrics.
\emph{Behavior Research Methods}, 1-23.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
