% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optmism.R
\docType{data}
\name{optmism}
\alias{optmism}
\title{Optmism Data}
\format{A 282x10 response matrix}
\usage{
data(optmism)
}
\description{
A response matrix (n = 282) containing responses to 10 items of the Revised Life Orientation Test (LOT-R), developed by Scheier, Carver, & Bridges (1994).
}
\examples{
data("optmism")

}
\references{
Scheier, M. F., Carver, C. S., & Bridges, M. W. (1994).
Distinguishing optimism from neuroticism (and trait anxiety, self-mastery, and self-esteem): a reevaluation of the Life Orientation Test.
\emph{Journal of personality and social psychology}, \emph{67(6)}, 1063-1078.
doi: \href{https://doi.org/10.1037//0022-3514.67.6.1063}{J Pers Soc Psychol}
}
\keyword{datasets}
