% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.EGA.R
\name{plot.EGA}
\alias{plot.EGA}
\title{Plot method for EGA objects.}
\usage{
\method{plot}{EGA}(object, layout = "spring", vsize = 6, ...)
}
\arguments{
\item{object}{An EGA object}
}
\description{
\code{plot} Plots the EGA result using \code{\link{qgraph}}
}
\examples{
ega.wmt <- EGA(data = wmt2[,7:24], plot.EGA = TRUE)
summary(ega.wmt)
plot(ega.wmt)

\dontrun{
plot(EGA)
}
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
\seealso{
\code{\link{EGA}} to estimate the number of dimensions of an instrument using EGA and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}

