% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.NetLoads.R
\name{plot.NetLoads}
\alias{plot.NetLoads}
\title{Plot for net.loads objects of \code{\link[EGAnet]{net.loads}} results}
\usage{
\method{plot}{NetLoads}(x, ...)
}
\arguments{
\item{x}{An \code{\link[EGAnet]{net.loads}} object}

\item{...}{potentially further arguments (\strong{unused currently})}
}
\description{
Returns a plot of the net.loads results of \code{\link[EGAnet]{net.loads}} results
}
\examples{
# Load data
wmt <- wmt2[,7:24]

\dontrun{
# Estimate EGA
ega.wmt <- EGA(wmt)

}

# Network loadings
nloads <- net.loads(ega.wmt)

# Plot
plot(nloads)

}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
