% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/prime.num.R
\docType{data}
\name{prime.num}
\alias{prime.num}
\title{Prime Numbers through 100,000}
\format{
A 1185x24 response matrix

A 1185x24 response matrix
}
\usage{
data(prime.num)

data(prime.num)
}
\description{
Numeric vector of primes generated from the primes package. Used in
the function \code{[EGAnet]{ergoInfo}}. Not for general use

Numeric vector of primes generated from the primes package. Used in
the function \code{[EGAnet]{ergoInfo}}. Not for general use
}
\examples{
data("prime.num")

data("prime.num")

}
\keyword{datasets}
