% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dynEGA.Groups.R
\name{print.dynEGA.Groups}
\alias{print.dynEGA.Groups}
\title{Print method for \code{\link[EGAnet]{dynEGA}} objects (Group Effects)}
\usage{
\method{print}{dynEGA.Groups}(x, ...)
}
\arguments{
\item{x}{An \code{\link[EGAnet]{dynEGA}} objects (Group Effects)}

\item{...}{potentially further arguments (\strong{unused currently})}
}
\description{
Returns a summary of the \code{\link[EGAnet]{dynEGA}} objects (Group Effects)
}
\examples{
\dontrun{
# Estimate dynEGA
dyn.group <- dynEGA(data = sim.dynEGA, n.embed = 5, tau = 1,
delta = 1, id = 21, group = 22, use.derivatives = 1,
effects = "group", model = "glasso")

#Print dynEGA results
print(dyn.group)
}

}
\seealso{
\code{\link[EGAnet]{dynEGA}} to estimate the number of dimensions in multivariate time series using dynEGA.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
