% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dynEGA.R
\name{print.dynEGA}
\alias{print.dynEGA}
\title{Print method for \code{\link[EGAnet]{dynEGA}} objects}
\usage{
\method{print}{dynEGA}(x, ...)
}
\arguments{
\item{x}{An \code{\link[EGAnet]{dynEGA}} objects}

\item{...}{potentially further arguments (\strong{unused currently})}
}
\description{
Returns a summary of the \code{\link[EGAnet]{dynEGA}} objects
}
\examples{
\dontrun{
# Estimate dynEGA
dyn.random <- dynEGA(data = sim.dynEGA, n.embed = 5, tau = 1,
delta = 1, id = 21, group = 22, use.derivatives = 1,
level = "population", model = "glasso")

#Print dynEGA results
print(dyn.random)
}

}
\seealso{
\code{\link[EGAnet]{dynEGA}} to estimate the number of dimensions in multivariate time series using dynEGA.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
