% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{prints}
\alias{prints}
\alias{print.hierEGA}
\title{S3Methods for Printing}
\usage{
\method{print}{hierEGA}(x, ...)
}
\arguments{
\item{x}{Object from \code{EGAnet} package}

\item{...}{Additional arguments}
}
\value{
Prints \code{EGAnet} object
}
\description{
Prints for \code{EGAnet} objects
}
\author{
Hudson Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
