% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyEGA.R
\name{shinyEGA}
\alias{shinyEGA}
\title{Shiny App for \code{\link{EGAnet}}}
\usage{
shinyEGA()
}
\value{
A list called \code{resultShiny} containing:

\item{data}{The data imported into \code{\link[EGAnet]{shinyEGA}}}

\item{EGA}{The output generated from \code{\link[EGAnet]{EGA}}}
}
\description{
An interactive Shiny application for running \code{\link{EGAnet}} analysis
}
\examples{
if(interactive())
{shinyEGA()}

}
\author{
Hudson Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
