% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed.louvain.R
\name{signed.louvain}
\alias{signed.louvain}
\title{Computes the Signed Louvain Community Detection Algorithm}
\usage{
signed.louvain(network)
}
\arguments{
\item{network}{Matrix or data frame.
A symmetric matrix representing a network}
}
\value{
Returns a list:

\item{communities}{Multi-level community matrix}

\item{modularities}{Modularity values for each level}
}
\description{
Computes the Signed Louvain Community Detection Algorithm
}
\examples{
# Load data
wmt <- wmt2[,7:24]

# Estimate network
network <- EBICglasso.qgraph(data = wmt)

# Estimate signed Louvain
signed.louvain(network)

}
\references{
Blondel, V. D., Guillaume, J.-L., Lambiotte, R., & Lefebvre, E. (2008).
Fast unfolding of communities in large networks.
\emph{Journal of Statistical Mechanics: Theory and Experiment}, \emph{2008}(10), P10008.

Gomez, S., Jensen, P., & Arenas, A. (2009).
Analysis of community structure in networks of correlated data.
\emph{Physical Review E}, \emph{80}(1), 016114.
}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com> with assistance from GPT-4
}
