% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_louvain.R
\name{signed_louvain}
\alias{signed_louvain}
\title{Computes the signed Louvain community detection algorithm}
\usage{
signed_louvain(network)
}
\arguments{
\item{network}{Matrix or data frame.
A symmetric matrix representing a network}
}
\value{
Returns a list:

\item{communities}{Multi-level community matrix}

\item{modularities}{Modularity values for each level}
}
\description{
Computes the signed Louvain community detection algorithm
}
\examples{
# Obtain data
wmt <- wmt2[,7:24]

\dontrun{
# Estimate EGA
ega.wmt <- EGA(
  data = wmt,
  plot.EGA = FALSE # No plot for CRAN checks
)}

# Estimate signed Louvain
signed_louvain(ega.wmt$network)

}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com> with assistance from GPT-4
}
