% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simEGM.R
\name{simEGM}
\alias{simEGM}
\title{Simulate data following a Exploratory Graph Model (\code{\link[EGAnet]{EGM}})}
\usage{
simEGM(
  communities,
  variables,
  loadings = c("small", "moderate", "large"),
  cross.loadings = 0.01,
  correlations = c("none", "small", "moderate", "large", "very large"),
  sample.size,
  p.in = 0.95,
  p.out = 0.9,
  max.iterations = 1000
)
}
\arguments{
\item{communities}{Numeric (length = 1).
Number of communities to generate}

\item{variables}{Numeric vector (length = 1 or \code{communities}).
Number of variables per community}

\item{loadings}{Character (length = 1).
Magnitude of the assigned network loadings.
Available options (based on revised network loadings with \code{scaling = 2}):

\itemize{

\item \code{"small"} --- 0.225 (approximates 0.20)

\item \code{"moderate"} --- 0.35 (approximates 0.35)

\item \code{"large"} --- 0.50 (approximates 0.50)

}

Uses \code{runif(n, min = value - 0.025, max = value + 0.025)} for some jitter in the loadings}

\item{cross.loadings}{Numeric (length = 1).
Standard deviation of a normal distribution with a mean of zero (\code{n, mean = 0, sd = value}).
Defaults to \code{0.01}}

\item{correlations}{Character (length = 1).
Magnitude of the community correlations.
Available options:

\itemize{

\item \code{"none"} --- 0.00

\item \code{"small"} --- 0.10

\item \code{"moderate"} --- 0.30

\item \code{"large"} --- 0.50

\item \code{"very large"} --- 0.70

}

Uses \code{rnorm(n, mean = value, sd = 0.03)} for some jitter in the correlations}

\item{sample.size}{Numeric (length = 1).
Number of observations to generate}

\item{p.in}{Numeric (length = 1).
Sets the probability of retaining an edge \emph{within} communities.
Single values are applied to all communities.
Defaults to \code{0.95}}

\item{p.out}{Numeric (length = 1 or \code{communities}).
Sets the probability of retaining an edge \emph{between} communities.
Single values are applied to all communities.
Defaults to \code{0.90}}

\item{max.iterations}{Numeric (length = 1).
Number of iterations to attempt to get convergence before erroring out.
Defaults to \code{1000}}
}
\description{
Function to simulate data based on \code{\link[EGAnet]{EGM}}
}
\examples{
\dontrun{
# Estimate EGM
EGM_data <- simEGM(
  communities = 2, variables = 6,
  loadings = "moderate", correlations = "moderate",
  sample.size = 1000
)}

}
\author{
Hudson F. Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
