% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsamples.R
\name{subsamples}
\alias{subsamples}
\title{Applying EGA to n random subsamples.}
\usage{
subsamples(data, n, ncores)
}
\arguments{
\item{data}{A data.frame object.}

\item{n}{Numeric. Number of random subsamples}

\item{ncores}{Numeric. Number of cores to use in parallel.}
}
\description{
\code{subsamples} Applies EGA to n random subsamples of the original data.
}
\examples{
\dontrun{
wmt.subsample <- subsamples(data = wmt2[,7:24], n = 100, ncores = 4)
}
\dontrun{
EGA.subsamples()
}
}
\seealso{
\code{\link{EGA}} to estimate the number of dimensions of an instrument using EGA, \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis, \code{\link{shuffle}} to generate n
estimations of the number of dimensions in surrogate samples of the original dataset.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
