% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.EGA.R
\name{summary.EGA}
\alias{summary.EGA}
\title{Summary for \code{\link[EGAnet]{EGA}} objects}
\usage{
\method{summary}{EGA}(object, ...)
}
\arguments{
\item{object}{An \code{\link[EGAnet]{EGA}} object}

\item{...}{potentially further arguments (\strong{unused currently})}
}
\description{
Returns a summary of the \code{\link[EGAnet]{EGA}} results
}
\examples{
\dontrun{
# Estimate EGA
ega.wmt <- EGA(data = wmt2[,7:24], plot.EGA = TRUE)

}

#Summary of EGA reults
summary(ega.wmt)

}
\seealso{
\code{\link[EGAnet]{EGA}} to estimate the number of dimensions of an instrument using EGA
and \code{\link[EGAnet]{CFA}} to verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
