% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.NetLoads.R
\name{summary.NetLoads}
\alias{summary.NetLoads}
\title{Summary for net.loads objects of \code{\link[EGAnet]{net.loads}} results}
\usage{
\method{summary}{NetLoads}(object, ...)
}
\arguments{
\item{object}{An \code{\link[EGAnet]{net.loads}} object}

\item{...}{potentially further arguments (\strong{unused currently})}
}
\description{
Returns a summary of the net.loads results of \code{\link[EGAnet]{net.loads}} results
}
\examples{
# Load data
wmt <- wmt2[,7:24]

\dontrun{
# Estimate EGA
ega.wmt <- EGA(wmt)

}

# Network loadings
nloads <- net.loads(ega.wmt)

# Summary
summary(nloads)

}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
