% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarys.R
\name{summarys}
\alias{summarys}
\alias{summary.CFA}
\alias{summary.dynEGA}
\alias{summary.dynEGA.Groups}
\alias{summary.dynEGA.Individuals}
\alias{summary.EGA}
\alias{summary.NetLoads}
\title{S3Methods for Summaries}
\usage{
\method{summary}{CFA}(object, ...)

\method{summary}{dynEGA}(object, ...)

\method{summary}{dynEGA.Groups}(object, ...)

\method{summary}{dynEGA.Individuals}(object, ...)

\method{summary}{EGA}(object, ...)

\method{summary}{NetLoads}(object, ...)
}
\arguments{
\item{object}{Object from \code{EGAnet} package}

\item{...}{Additional arguments}
}
\value{
Summarizes \code{EGAnet} object
}
\description{
Summaries for \code{EGAnet} objects
}
\author{
Hudson Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
