% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate.R
\name{surrogate}
\alias{surrogate}
\title{Surrogate analysis for EGA.}
\usage{
surrogate(data, n, ncores, plot.surrogate = TRUE)
}
\arguments{
\item{data}{A data.frame object.}

\item{n}{Numeric. Number of estimates.}

\item{ncores}{Numeric. Number of cores to use in parallel.}

\item{plot.surrogate}{Logic. If true, returns a plot with the surrogate analysis results.}
}
\description{
\code{surrogate} Compares the median number of dimensions (and it's 95% confidence intervals) estimated in n shuffled versions of the
original dataset and in n random subsamples (\code{\link{subsamples}}) of the data.
}
\examples{
\dontrun{
surrogate.analysis <- surrogate(data = wmt2[,7:24], n = 100, ncores = 4)
}
}
\seealso{
\code{\link{subsamples}} to estimate the number of dimensions via EGA in n random subsamples of the original data and \code{\link{shuffle}}
to generate n estimations of the number of dimensions in shuffled versions of the original dataset
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
