% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tefi.R
\name{tefi}
\alias{tefi}
\title{Total Entropy Fit Index using Von Neumman's entropy (Quantum Information Theory) for correlation matrices}
\usage{
tefi(data, structure = NULL)
}
\arguments{
\item{data}{A matrix, data frame, or correlation matrix}

\item{structure}{A vector representing the structure (numbers or labels for each item).
Can be theoretical factors or the structure detected by \code{\link{EGA}}. For high-order and correlated bifactor structures,
\code{structure} should be a list containing:

\itemize{

\item{\code{lower_order}}
{A vector representing the first-order structure (numbers or labels for each item in each first-order factor or community).}

\item{\code{higher_order}}
{A vector representing the second-order structure (numbers or labels for each item in each second-order factor or community).}

}}
}
\value{
Returns a list containing:

\item{VN.Entropy.Fit}{The Entropy Fit Index using Von Neumman's entropy}

\item{Total.Correlation}{The total correlation of the dataset}

\item{Average.Entropy}{The average entropy of the dataset}
}
\description{
Computes the fit (TEFI) of a dimensionality structure using Von Neumman's entropy when the input is a correlation matrix.
Lower values suggest better fit of a structure to the data. For high-order and correlated bifactor structures (or \code{\link{hierEGA}} objects), the Generalized
Total Entropy Fit index is computed.
}
\examples{
# Load data
wmt <- wmt2[,7:24]

# Estimate EGA model
ega.wmt <- EGA(
  data = wmt, model = "glasso",
  plot.EGA = FALSE # no plot for CRAN checks
)

# Compute entropy indices
tefi(data = ega.wmt$correlation, structure = ega.wmt$wc)

# Load data
data <- optimism

# Not run: 
# hierEGA example
 opt.hier<- hierEGA(
 data = optimism,
 algorithm = "louvain")
 
 # Create a list with the lower and higher order structures:
 hier.structure <- vector("list")
 hier.structure$lower_order <- opt.hier$dim.variables$lower
 hier.structure$higher_order <- opt.hier$dim.variables$higher
 
 # Compute the Generalized Total Entropy Fit Index
 gen.tefi.opt <- tefi(opt.hier$lower_order$correlation, structure = hier.structure)
 

}
\references{
Golino, H., Moulder, R. G., Shi, D., Christensen, A. P., Garrido, L. E., Nieto, M. D., Nesselroade, J., Sadana, R., Thiyagarajan, J. A., & Boker, S. M. (2020).
Entropy fit indices: New fit measures for assessing the structure and dimensionality of multiple latent variables.
\emph{Multivariate Behavioral Research}.
}
\author{
Hudson Golino <hfg9s at virginia.edu>, Alexander P. Christensen <alexpaulchristensen@gmail.com>, and Robert Moulder <rgm4fd@virginia.edu>
}
