% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalCor.R
\name{totalCor}
\alias{totalCor}
\title{Total Correlation}
\usage{
totalCor(data)
}
\arguments{
\item{data}{Matrix or data frame.
Variables to be used in the analysis}
}
\value{
Returns a list containing:

\item{Ind.Entropies}{Individual entropies for each variable}

\item{Joint.Entropy}{The joint entropy of the dataset}

\item{Total.Cor}{The total correlation of the dataset}
}
\description{
Computes the total correlation of a dataset
}
\examples{
# Compute total correlation
totalCor(wmt2[,7:24])

}
\references{
Watanabe, S. (1960).
Information theoretical analysis of multivariate correlation.
\emph{IBM Journal of Research and Development} \emph{4}, 66-82.

# Implementation
Felix, L. M., Mansur-Alves, M., Teles, M., Jamison, L., & Golino, H. (2021).
Longitudinal impact and effects of booster sessions in a cognitive training program for healthy older adults.
\emph{Archives of Gerontology and Geriatrics}, \emph{94}, 104337.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
