% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UVA.R
\name{UVA}
\alias{UVA}
\title{Unique Variable Analysis}
\usage{
UVA(
  data = NULL,
  network = NULL,
  n = NULL,
  key = NULL,
  cut.off = 0.25,
  reduce = TRUE,
  reduce.method = c("latent", "mean", "remove", "sum"),
  auto = TRUE,
  label.latent = FALSE,
  verbose = FALSE,
  uva.method = c("MBR", "EJP"),
  ...
)
}
\arguments{
\item{data}{Matrix, data frame, or symmetric matrix.
Input can either be data or a correlation matrix}

\item{network}{Symmetric matrix.
Input for a symmetric network matrix.

If both \code{data} and \code{network} are provided,
then \code{UVA} will proceed to use the input network
with the data (rather than estimating a network)}

\item{n}{Numeric vector (length = 1).
If input in \code{data} is a correlation matrix, 
then sample size is required.
Defaults to \code{NULL}}

\item{key}{Character vector (length = \code{ncol(data)}).
Item key for labeling items}

\item{cut.off}{Numeric vector (length = 1).
Must be values between \code{0} and \code{1}.
Defaults to \code{0.25}}

\item{reduce}{Logical (length = 1).
Whether redundancies should be reduced in data.
Defaults to \code{TRUE}}

\item{reduce.method}{Character.
Method to reduce redundancies:

\itemize{

\item{\code{"latent"}}
{Computes latent variables when there are three or more
redundant variables. Computes sum otherwise.
}

\item{\code{"remove"}}
{Removes all but one variable from a set of redundant variables
}

\item{\code{"sum"}}
{Computes sum of redundant variables
}

}}

\item{auto}{Logical (length = 1).
Whether \code{reduce} should occur automatically using different rules
depending on reduction method:

\itemize{

\item{\code{"latent"}}
{Computes latent variables when there are three or more
redundant variables. Computes sum otherwise.
}

\item{\code{"remove"}}
{Removes all but one variable from a set of redundant variables.
Keeps the variable with the highest variable-total correlation using
the correlation between each variable and the sum of all other
variables with the target variable removed. For ties and two variables,
the variable with the largest standard deviation is kept.
}

\item{\code{"sum"}}
{Computes sum of redundant variables
}

}}

\item{label.latent}{Boolean (length = 1).
When \code{reduce.method = "latent"}, should
latent variables be labelled?
Defaults to \code{FALSE}.
Set to \code{TRUE} to type your own labels}

\item{...}{Additional arguments.
Arguments that should be passed onto old versions of \code{UVA}
(use the \code{EGAnet.version} argument to set the version),
\code{\link[EGAnet]{EGA}}, and \code{\link[lavaan]{cfa}}}

\item{EGAnet.version}{Character (length = 1).
\code{\link{EGAnet}} version used to perform previous
\code{UVA}.
Defaults to current version.
Set version to previous version to reproduce results
from an older version}
}
\description{
Identifies redundant variables in a multivariate dataset
using the \code{\link[qgraph]{EBICglasso}} network estimation method
and weighted topological overlap
(see Christensen, Garrido, & Golino, 2020 for more details)
}
\examples{
# Select Five Factor Model personality items only
idx <- na.omit(match(gsub("-", "", unlist(psychTools::spi.keys[1:5])), colnames(psychTools::spi)))
items <- psychTools::spi[,idx]

# Change names in redundancy output to each item's description
key.ind <- match(colnames(items), as.character(psychTools::spi.dictionary$item_id))
key <- as.character(psychTools::spi.dictionary$item[key.ind])

# Results with no reduction
no_reduce_uva <- UVA(
  data = items, key = key, reduce = FALSE
)

# Show summary
summary(no_reduce_uva)

\dontrun{
# Results with automatic reduction
reduced_uva <- UVA(
  data = items, key = key, reduce = TRUE,
  reduce.method = "latent" 
)}

}
\references{
# Simulation using UVA
Christensen, A. P., Garrido, L. E., & Golino, H. (under review).
Unique variable analysis: A network psychometrics method to detect local dependence.
\emph{PsyArXiv}.

# Implementation of UVA (formally node.redundant)
Christensen, A. P., Golino, H., & Silvia, P. J. (2020).
A psychometric network perspective on the validity and validation of personality trait questionnaires.
\emph{European Journal of Personality}, \emph{34}(6), 1095-1108.

# wTO measure
Nowick, K., Gernat, T., Almaas, E., & Stubbs, L. (2009).
Differences in human and chimpanzee gene expression patterns define an evolving network of transcription factors in brain.
\emph{Proceedings of the National Academy of Sciences}, \emph{106}, 22358-22363.

# Selection of CFA Estimator
Rhemtulla, M., Brosseau-Liard, P. E., & Savalei, V. (2012).
When can categorical variables be treated as continuous? A comparison of robust continuous and categorical SEM estimation methods under suboptimal conditions.
\emph{Psychological Methods}, \emph{17}(3), 354-373.
}
