% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmt2.R
\docType{data}
\name{wmt2}
\alias{wmt2}
\title{WMT-2 Data}
\format{A 1185x24 response matrix}
\usage{
data(wmt2)
}
\description{
A response matrix (n = 1185) of the Wiener Matrizen-Test 2 (WMT-2).
}
\examples{
data("wmt2")

}
\keyword{datasets}
