% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.correlate.R
\name{auto.correlate}
\alias{auto.correlate}
\title{Automatic correlations based on \code{\link[qgraph]{cor_auto}}}
\usage{
auto.correlate(
  data,
  ordinal.categories = 7,
  forcePD = TRUE,
  missing = "pairwise",
  verbose = FALSE
)
}
\arguments{
\item{data}{Matrix or data frame.
Should consist only of variables that are desired to be correlated}

\item{ordinal.categories}{Numeric (length = 1).
Number of categories before a variable is considered continuous.
Defaults to \code{7} categories before \code{8} is considered continuous}

\item{forcePD}{Boolean (length = 1).
Whether positive definite matrix should be enforced.
Defaults to \code{TRUE}}

\item{missing}{Character (length = 1).
Corresponds to \code{\link[lavaan]{lavCor}}'s \code{missing} argument}

\item{verbose}{Boolean (length = 1).
Whether messages should be printed.
Defaults to \code{FALSE}}
}
\description{
This wrapper is based on \code{\link[qgraph]{cor_auto}}. There
are some minor adjustments that make this function simpler and to
function within EGAnet as desired. \code{NA} values are not treated
as categories
}
\examples{
# Load data
wmt <- wmt2[,7:24]

# Obtain correlations
wmt_corr <- auto.correlate(wmt)

}
\references{
Sacha Epskamp, Angelique O. J. Cramer, Lourens J. Waldorp, Verena D. Schmittmann, Denny Borsboom (2012). qgraph: Network Visualizations of Relationships in Psychometric Data. Journal of Statistical Software, 48(4), 1-18.

Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. URL http://www.jstatsoft.org/v48/i02/.

Douglas Bates and Martin Maechler (2014). Matrix: Sparse and Dense Matrix Classes and Methods. R package version 1.1-3. http://CRAN.R-project.org/package=Matrix
}
\author{
Sacha Epskamp <mail@sachaepskamp.com> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
