% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootEGA.R
\name{bootEGA}
\alias{bootEGA}
\title{Investigates the stability of EGA's estimation via bootstrap.}
\usage{
bootEGA(data, n, medianStructure = TRUE, plot.MedianStructure = TRUE,
  ncores = 4)
}
\arguments{
\item{data}{A dataframe with the variables to be used in the analysis}

\item{n}{An integer value representing the number of bootstraps}

\item{medianStructure}{Logical. If true, returns the typical network of partial correlations (estimated via graphical lasso), which is the median of all pairwise correlations over the n bootstraps, and estimates its dimensions.}

\item{plot.MedianStructure}{Logical. If true, returns a plot of the typical network (partial correlations), which is the median of all pairwise correlations over the n bootstraps, and its estimated dimensions.}

\item{ncores}{Number of cores to use in computing results. Set to 1 to not use parallel computing.}
}
\description{
\code{bootEGA} Estimates the number of dimensions of n bootstraps from the empirical correlation matrix,
 and returns a typical network (i.e. the network formed by the median pairwise partial correlations over the n bootstraps) and its dimensionality.
}
\examples{
boot.wmt <- bootEGA(data = wmt2[,7:24], n = 500, medianStructure = TRUE, plot.MedianStructure = TRUE)
boot.intwl <- bootEGA(data = intelligenceBattery[,8:66], n = 500, medianStructure = TRUE, plot.MedianStructure = TRUE, ncores = 4)

\dontrun{
bootEGA(a)
}
}
\seealso{
\code{\link{EGA}} to estimate the number of dimensions of an instrument using EGA and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
