% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootEGA.R
\name{bootEGA}
\alias{bootEGA}
\title{Investigates the stability of EGA's estimation via bootstrap.}
\usage{
bootEGA(data, n, typicalStructure = TRUE, plot.typicalStructure = TRUE,
  ncores = 4, model = c("glasso", "TMFG"), type = c("parametric",
  "resampling"))
}
\arguments{
\item{data}{A dataframe with the variables to be used in the analysis}

\item{n}{An integer value representing the number of bootstraps}

\item{typicalStructure}{Logical. If true, returns the typical network of partial correlations (estimated via graphical lasso or via TMFG) and estimates its dimensions. The "typical network" is the median of all pairwise correlations over the n bootstraps.}

\item{plot.typicalStructure}{Logical. If true, returns a plot of the typical network (partial correlations), which is the median of all pairwise correlations over the n bootstraps, and its estimated dimensions.}

\item{ncores}{Number of cores to use in computing results. Set to 1 to not use parallel computing.}

\item{model}{A string indicating the method to use. Current options are:
-\code{glasso}:
{Gaussian Markov random field estimation using graphical LASSO with extended Bayesian information criterion to select optimal regularization parameter. Using \code{\link[qgraph]{EBICglasso}} from the qgraph package version 1.4.4.}
\code{TMFG}:
{Estimates a Triangulated Maximally Filtered Graph, using the function \code{TMFG} of the NetworkToolbox package}}

\item{type}{A string indicating the type of bootstrap to use. Current options are:
-\code{parametric}:
{Generates n new datasets (multivariate normal random distributions) based on the
original dataset, via the \code{\link[mvtnorm]{rmvnorm}} function of the mvtnorm package}.
-\code{resampling}:
{Generates n random subsamples of the original data.}}
}
\description{
\code{bootEGA} Estimates the number of dimensions of n bootstraps from the empirical correlation matrix,
 and returns a typical network (i.e. the network formed by the median or mean pairwise correlations over the n bootstraps) and its dimensionality.
}
\examples{
\dontrun{
boot.wmt <- bootEGA(data = wmt2[,7:24], n = 500, typicalStructure = TRUE,
plot.typicalStructure = TRUE, model = "glasso", type = "parametric", ncores = 4)
boot.intwl <- bootEGA(data = intelligenceBattery[,8:66], n = 500, typicalStructure = TRUE,
plot.typicalStructure = TRUE, model = "glasso", type = "parametric", ncores = 4)
}
}
\seealso{
\code{\link{EGA}} to estimate the number of dimensions of an instrument using EGA and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu> and Alexander Christensen <alexpaulchristensen@gmail.com>
}
