% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.EGA.plots.R
\name{compare.EGA.plots}
\alias{compare.EGA.plots}
\title{Visual Compares \code{\link{EGAnet}} plots}
\usage{
compare.EGA.plots(..., rows, columns)
}
\arguments{
\item{...}{\code{\link{EGAnet}} objects}

\item{rows}{Numeric.
Number of rows to spread plots across}

\item{columns}{Numeric.
Number of columns to spread plots down}
}
\value{
Visual comparison of \code{\link{EGAnet}} objects
}
\description{
Identifies redundant nodes in the network based on several
measures. Computes the weighted topological overlap between
each node and every other node in the network. The weighted topological
overlap is implemented using the method from Nowick et al. (2009; see references)
and the function \link[wTO]{wTO} from the wTO package.
}
\examples{
# obtain SAPA items
items <- psychTools::spi[,c(11:20)]

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
