% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGA.R
\name{EGA}
\alias{EGA}
\title{Apply the Exploratory Graph Analysis technique}
\usage{
EGA(data, model = c("glasso", "TMFG"), plot.EGA = TRUE, n = NULL,
  steps = 4)
}
\arguments{
\item{data}{A dataframe with the variables to be used in the analysis, or a correlation matrix. If the data used is a correlation matrix, the arguments *matrix* and *n* will need to be specified.}

\item{model}{A string indicating the method to use. Current options are:
-\code{glasso}:
{Gaussian graphical model estimation using graphical LASSO with extended Bayesian information criterion to select optimal regularization parameter (default method). Using \code{\link[qgraph]{EBICglasso}} from the \code{\link[qgraph]{qgraph}} package.}
\code{TMFG}:
{Estimates a Triangulated Maximally Filtered Graph, using the function \code{\link[NetworkToolbox]{TMFG}} of the \code{\link[NetworkToolbox]{NetworkToolbox}} package}}

\item{plot.EGA}{Logical. If TRUE, returns a plot of the network and its estimated dimensions.}

\item{n}{Integer. Sample size, if the data provided is a correlation matrix.}

\item{steps}{Number of steps to be used in walktrap algorithm.
Defaults to 4}
}
\description{
\code{EGA} Estimates the number of dimensions of a given dataset/instrument using graphical lasso or the TMFG method and a random walk algorithm. The glasso regularization parameter
is set via EBIC.
}
\examples{
ega.wmt <- EGA(data = wmt2[,7:24], model = "glasso", plot.EGA = TRUE)
summary(ega.wmt)
plot(ega.wmt)

ega.intel <- EGA(data = intelligenceBattery[,8:66], model = "glasso", plot.EGA = TRUE)
summary(ega.intel)
plot(ega.intel)

\dontrun{
EGA(a)
}
}
\references{
Golino, H. F., & Epskamp, S. (2017).
Exploratory graph analysis: A new approach for estimating the number of dimensions in psychological research.
\emph{PloS one}, \emph{12(6)}, e0174035..
doi: \href{https://doi.org/10.1371/journal.pone.0174035}{journal.pone.0174035}

Golino, H. F., & Demetriou, A. (2017).
Estimating the dimensionality of intelligence like data using Exploratory Graph Analysis.
\emph{Intelligence}, \emph{62}, 54-70.
doi: \href{https://doi.org/10.1016/j.intell.2017.02.007}{j.intell.2017.02.007}

Golino, H., Shi, D., Garrido, L. E., Christensen, A. P., Nieto, M. D., Sadana, R., & Thiyagarajan, J. A. (2018).
Investigating the performance of Exploratory Graph Analysis and traditional techniques to identify the number of latent factors: A simulation and tutorial.
\emph{psyarxiv}.
doi: \href{https://psyarxiv.com/gzcre/}{10.31234/osf.io/gzcre}
}
\seealso{
\code{\link{bootEGA}} to investigate the stability of EGA's estimation via bootstrap and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
