% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGA.R
\name{EGA}
\alias{EGA}
\title{Apply the Exploratory Graph Analysis technique}
\usage{
EGA(data, plot.EGA = TRUE, matrix = FALSE, n = NULL)
}
\arguments{
\item{data}{A dataframe with the variables to be used in the analysis, or a correlation matrix. If the data used is a correlation matrix, the arguments *matrix* and *n* will need to be specified.}

\item{plot.EGA}{Logical. If TRUE, returns a plot of the network of partial correlations estimated via graphical lasso and its estimated dimensions.}

\item{matrix}{Logical. If TRUE, will treat the data as a correlation matrix, and a *n* (sample size) will need to be specified. Default to FALSE.}

\item{n}{Integer. Sample size, if the data provided is a correlation matrix.}
}
\description{
\code{EGA} Estimates the number of dimensions of a given dataset/instrument using graphical lasso and a random walk algorithm. The glasso regularization parameter
is set via EBIC.
}
\examples{
ega.wmt <- EGA(data = wmt2[,7:24], plot.EGA = TRUE)
summary(ega.wmt)
plot(ega.wmt)

ega.intel <- EGA(data = intelligenceBattery[,8:66])
summary(ega.intel)
plot(ega.intel)

\dontrun{
EGA(a)
}
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
\seealso{
\code{\link{bootEGA}} to investigate the stability of EGA's estimation via bootstrap and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}

