% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega.wmt.R
\docType{data}
\name{ega.wmt}
\alias{ega.wmt}
\title{\code{\link[EGAnet]{EGA}} Network of \code{\link[EGAnet]{wmt2}}Data}
\format{
A list with 8 objects (see \strong{Value} in \code{\link[EGAnet]{EGA}})
}
\usage{
data(ega.wmt)
}
\description{
\code{\link[EGAnet]{EGA}} results from \code{ega.wmt <- EGA(wmt2[,7:24])}
for the Wiener Matrizen-Test (WMT-2)
}
\examples{
data("ega.wmt")
}
\keyword{datasets}
