% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega.wmt.R
\docType{data}
\name{ega.wmt}
\alias{ega.wmt}
\title{\code{\link[EGAnet]{EGA}} Network of \code{\link[EGAnet]{wmt2}}Data}
\format{A 17 x 17 adjacency matrix}
\usage{
data(ega.wmt)
}
\description{
An \code{\link[EGAnet]{EGA}} using the \code{"glasso"} model of the
Wiener Matrizen-Test 2 (WMT-2)
}
\examples{
data("ega.wmt")

}
\keyword{datasets}
