% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyFit.R
\name{entropyFit}
\alias{entropyFit}
\title{Fit of Factor Model using Empirical Entropy}
\usage{
entropyFit(data, structure)
}
\arguments{
\item{data}{A dataset}

\item{structure}{A vector representing the structure (numbers or labels for each item).
Can be theoretical factors or the structure detected by \code{\link{EGA}}}
}
\value{
Returns the values of entropy per factor, the mean empirical entropy (i.e., the average entropy across factors),
the joint entropy (i.e. joint entropy for all factors), a modified empirical entropy per factor (that uses the frequencies of the sumscores),
a mean modified empirical entropy. It also returns an adjusted entropy (ratio between the mean empirical entropy and
the joint entropy) and an adjusted entropy varying from 0 to 1.
}
\description{
Compute compute the entropy fit of a given structure.
Lower values of the adjusted entropy fit and the adjusted entropy fit
ranging from 0 to 1 suggest better fit with the data
}
\examples{
ega.wmt <- EGA(data = wmt2[,7:24], model = "glasso")

entropyFit(data = wmt2[,7:24], structure = ega.wmt$wc)

}
\seealso{
\code{\link{EGA}} to estimate the number of dimensions of an instrument using EGA and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu> and Alexander Christensen <alexpaulchristensen@gmail.com>
}
