% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierEGA.R
\name{hierEGA}
\alias{hierEGA}
\title{Hierarchical \code{\link[EGAnet]{EGA}}}
\usage{
hierEGA(
  data,
  scores = c("factor", "network"),
  consensus_iter = 1000,
  uni.method = c("expand", "LE"),
  corr = c("cor_auto", "pearson", "spearman"),
  model = c("glasso", "TMFG"),
  model.args = list(),
  algorithm = c("walktrap", "louvain"),
  algorithm.args = list(),
  plot.EGA = TRUE,
  plot.type = c("GGally", "qgraph"),
  plot.args = list(),
  verbose = TRUE
)
}
\arguments{
\item{data}{Matrix or data frame.
Variables (down columns) only.
Does not accept correlation matrices}

\item{scores}{Character.
How should scores for the lower-order structure be estimated?
Defaults to \code{"network"} for network scores computed using
the \code{\link[EGAnet]{net.scores}} function.
Set to \code{"factor"} for factor scores computed using
\code{\link[psych]{fa}}. Factors are assumed to be correlated
using the \code{"oblimin"} rotation}

\item{consensus_iter}{Numeric.
Number of iterations to perform in consensus clustering
(see Lancichinetti & Fortunato, 2012).
Defaults to \code{1000}}

\item{uni.method}{Character.
What unidimensionality method should be used? 
Defaults to \code{"LE"}.
Current options are:

\itemize{

\item{\strong{\code{expand}}}
{Expands the correlation matrix with four variables correlated .50.
If number of dimension returns 2 or less in check, then the data 
are unidimensional; otherwise, regular EGA with no matrix
expansion is used. This is the method used in the Golino et al. (2020)
\emph{Psychological Methods} simulation.}

\item{\strong{\code{LE}}}
{Applies the leading eigenvalue algorithm (\code{\link[igraph]{cluster_leading_eigen}})
on the empirical correlation matrix. If the number of dimensions is 1,
then the leading eigenvalue solution is used; otherwise, regular EGA
is used. This is the final method used in the Christensen, Garrido,
and Golino (2021) simulation.}

}}

\item{corr}{Type of correlation matrix to compute. The default uses \code{\link[qgraph]{cor_auto}}.
Current options are:

\itemize{

\item{\strong{\code{cor_auto}}}
{Computes the correlation matrix using the \code{\link[qgraph]{cor_auto}} function from
\code{\link[qgraph]{qgraph}}}.

\item{\strong{\code{pearson}}}
{Computes Pearson's correlation coefficient using the pairwise complete observations via
the \code{\link[stats]{cor}}} function.

\item{\strong{\code{spearman}}}
{Computes Spearman's correlation coefficient using the pairwise complete observations via
the \code{\link[stats]{cor}}} function.
}}

\item{model}{Character.
A string indicating the method to use.
Defaults to \code{"glasso"}.
Current options are:

\itemize{

\item{\strong{\code{glasso}}}
{Estimates the Gaussian graphical model using graphical LASSO with
extended Bayesian information criterion to select optimal regularization parameter}

\item{\strong{\code{TMFG}}}
{Estimates a Triangulated Maximally Filtered Graph}

}}

\item{model.args}{List.
A list of additional arguments for \code{\link[EGAnet]{EBICglasso.qgraph}}
or \code{\link[NetworkToolbox]{TMFG}}}

\item{algorithm}{A string indicating the algorithm to use or a function from \code{\link{igraph}}
Defaults to \code{"louvain"}.
Current options are:

\itemize{

\item{\strong{\code{walktrap}}}
{Computes the Walktrap algorithm using \code{\link[igraph]{cluster_walktrap}}}

\item{\strong{\code{louvain}}}
{Computes the Louvain algorithm using \code{\link[igraph]{cluster_louvain}}}

}}

\item{algorithm.args}{List.
A list of additional arguments for \code{\link[igraph]{cluster_walktrap}}, \code{\link[igraph]{cluster_louvain}},
or some other community detection algorithm function (see examples)}

\item{plot.EGA}{Boolean.
If \code{TRUE}, returns a plot of the network and its estimated dimensions.
Defaults to \code{TRUE}}

\item{plot.type}{Character.
Plot system to use.
Current options are \code{\link[qgraph]{qgraph}} and \code{\link{GGally}}.
Defaults to \code{"GGally"}}

\item{plot.args}{List.
A list of additional arguments for the network plot.
For \code{plot.type = "qgraph"}:

\itemize{

\item{\strong{\code{vsize}}}
{Size of the nodes. Defaults to 6.}

}
For \code{plot.type = "GGally"} (see \code{\link[GGally]{ggnet2}} for
full list of arguments):

\itemize{

\item{\strong{\code{vsize}}}
{Size of the nodes. Defaults to 6.}

\item{\strong{\code{label.size}}}
{Size of the labels. Defaults to 5.}

\item{\strong{\code{alpha}}}
{The level of transparency of the nodes, which might be a single value or a vector of values. Defaults to 0.7.}

\item{\strong{\code{edge.alpha}}}
{The level of transparency of the edges, which might be a single value or a vector of values. Defaults to 0.4.}

 \item{\strong{\code{legend.names}}}
{A vector with names for each dimension}

\item{\strong{\code{color.palette}}}
{The color palette for the nodes. For custom colors,
enter HEX codes for each dimension in a vector.
See \code{\link[EGAnet]{color_palette_EGA}} for
more details and examples}

}}

\item{verbose}{Boolean.
Should network estimation parameters be printed?
Defaults to \code{TRUE}.
Set to \code{FALSE} for no print out}
}
\value{
Returns a list containing:

\item{lower}{Lower-order results from \code{\link[EGAnet]{EGA}}}

\item{lower_scores}{Estimated scores for the lower dimensions}

\item{lower_loadings}{Estaimted loadings for the lower dimensions}

\item{score_type}{Type of scores estimated and used in the higher-order \code{\link[EGAnet]{EGA}}}

\item{higher}{Higher-order results from \code{\link[EGAnet]{EGA}}}

\item{lower_plot}{Plot for the lower-order dimensions}

\item{higher_plot}{Plot for the higher-order dimensions}

\item{hier_plot}{Plot showing the lower-order and higher-order dimensions}
}
\description{
Estimates EGA using the lower-order solution of \code{\link[igraph]{cluster_louvain}}
to identify the lower-order dimensions and then uses factor or network loadings to estimate factor or
network scores to estimate the higher-order dimensions
}
\examples{
# Obtain example data
data <- optimism

\dontrun{
# hierEGA example
opt.res <- hierEGA(data = optimism, algorithm = "louvain")
}


}
\references{
Lancichinetti, A., & Fortunato, S. (2012).
Consensus clustering in complex networks.
\emph{Scientific Reports}, \emph{2}(1), 1-7.
}
\author{
Luis E. Garrido <garrido.luiseduardo@gmail.com>,
Alexander P. Christensen <alexpaulchristensen@gmail.com>, and
Hudson F. Golino <hfg9s at virginia.edu>
}
