% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoCluster.R
\name{infoCluster}
\alias{infoCluster}
\title{Variation of Information Hierarchical Clustering}
\usage{
infoCluster(dynEGA.object, plot.cluster = TRUE)
}
\arguments{
\item{dynEGA.object}{A \code{\link[EGAnet]{dynEGA}} or a
\code{\link[EGAnet]{dynEGA.pop.ind}} object that is used to match the arguments of the EII object.}

\item{plot.cluster}{Boolean.
Should plot of optimal and hierarchical clusters be output?
Defaults to \code{TRUE}.
Set to \code{FALSE} to not plot}
}
\value{
Returns a list containing:

\item{id_clusters}{A list containing the IDs for each cluster. 
\code{missing} contains IDs that were not able to be included due
to missing too many memberships (NA >= 0.50)}

\item{optimal}{Optimal clusters determine by *k*-means clustering from
\code{\link[factoextra]{fviz_nbclust}}}

\item{hierarchical}{Results of the hierarchical clustering based on the optimal
number of clusts. Reports results from \code{\link[factoextra]{hcut}}}

\item{plot}{Plot output from results}
}
\description{
Performs hierarchical k-means clustering using variation of information
of the community memberships derived from individuals in \code{\link[EGAnet]{dynEGA}}
}
\examples{

\dontrun{
\donttest{
# Perform dynEGA
dyn1 <- dynEGA.ind.pop(data = sim.dynEGA[,-c(22)], n.embed = 5, tau = 1,
                      delta = 1, id = 21, use.derivatives = 1,
                    model = "glasso", ncores = 2, corr = "pearson")

# Perform hierarchical clustering
clust1 <- infoCluster(
  dynEGA.object = dyn1,
  plot.cluster = TRUE
)
}}

}
\author{
Hudson Golino <hfg9s at virginia.edu> & Alexander P. Christensen <alexander.christensen at Vanderbilt.Edu>
}
