% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemConfirm.R
\name{itemConfirm}
\alias{itemConfirm}
\title{Replicability of items from EGA}
\usage{
itemConfirm(bootega.obj, confirm, plot.ic = TRUE)
}
\arguments{
\item{bootega.obj}{A \code{\link{bootEGA}} object}

\item{confirm}{A vector with community numbers or labels for each item}

\item{plot.ic}{Should the plot be produced?
Defaults to TRUE}
}
\value{
Returns a list containing:

\item{dim.confirm}{The proportion of times each dimension replicated
(i.e., all items were in the exact same dimension)}

\item{item.confirm}{The proporton of times each item replicated
within the defined dimension}

\item{dim.nmi.table}{Computes the Normalized Mutual Informaiton
for community replication. Larger numbers mean dimensions and
items are replicating closer to the input (\code{confirm})
dimensions (see Danon et al. 2005 for more details)}
}
\description{
Based on the \code{\link{bootEGA}} results, this function 
computes and plots the number of times an item (variable) is estimated
in the same factor/dimension as originaly estimated by EGA. The output
will also contain how often a dimension is replicated.
}
\examples{
\dontrun{
ega.wmt <- EGA(data = wmt2[,7:24], model = "glasso")

boot.wmt <- bootEGA(data = wmt2[,7:24], n = 100, typicalStructure = TRUE,
plot.typicalStructure = TRUE, model = "glasso",
type = "parametric", ncores = 4)

itemConfirm(boot.wmt, confirm = ega.wmt$wc)
}

}
\references{
Danon, L., Diaz-Guilera, A., Duch, J., & Arenas, A. (2005).
Comparing community structure identification.
\emph{Journal of Statistical Mechanics: Theory and Experiment}, \emph{9}, P09008.
doi:\href{https://doi.org/10.1088/1742-5468/2005/09/P09008}{10.1088/1742-5468/2005/09/P09008}
}
\seealso{
\code{\link{EGA}} to estimate the number of dimensions of an instrument using EGA and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu> and Alexander Christensen <alexpaulchristensen@gmail.com>
}
