% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCT.R
\name{LCT}
\alias{LCT}
\title{Loadings Comparison Test}
\usage{
LCT(data, n, iter = 100)
}
\arguments{
\item{data}{Matrix or data frame.
A dataframe with the variables to be used in the test or a correlation matrix.
If the data used is a correlation matrix, the argument \code{n} will need to be specified}

\item{n}{Integer.
Sample size (if the data provided is a correlation matrix)}

\item{iter}{Integer.
Number of replicate samples to be drawn from a multivariate
normal distribution (uses \code{\link[mvtnorm]{rmvnorm}}).
Defaults to \code{100}}
}
\value{
Returns a character string with the model suggested
from the three heuristics (see Details)
}
\description{
An algorithm to identify whether data were generated from a
random, factor, or network model using factor and network loadings.
The algorithm uses heuristics based on theory and simulation.
}
\examples{
\donttest{# Compute LCT
## Network model
LCT(data = wmt2[,7:24])

## Factor model
LCT(data = depression[,48:68])}

}
\references{
Christensen, A. P., & Golino, H. (2020).
Statistical equivalency of factor and network loadings.
\emph{PsyArXiv}.
doi:\href{https://doi.org/10.31234/osf.io/xakez}{10.31234/osf.io/xakez}
}
\author{
Hudson F. Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen at gmail.com>
}
