% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndim.R
\name{ndim}
\alias{ndim}
\title{Estimates the number of dimensions}
\usage{
ndim(data)
}
\arguments{
\item{data}{A dataframe}
}
\description{
A wrapper function that applies EGA to estimate the number
of dimensions for simulation studies, resampling, or surrogate analysis.
}
\examples{
#estimate number of dimensions in data
wmt.dim <- ndim(data = wmt2[,7:24])

}
\seealso{
\code{\link{EGA}} to estimate the number of dimensions of an instrument using EGA, \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis, \code{shuffle} to generate n
estimations of the number of dimensions in shuffled versions of the original dataset and \code{subsamples} to
apply EGA to n random subsamples of the original data.
}
