% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bootEGA.R
\name{plot.bootEGA}
\alias{plot.bootEGA}
\title{Plot method for bootEGA objects.}
\usage{
\method{plot}{bootEGA}(bootega.obj, vsize = 6, ...)
}
\arguments{
\item{bootega.obj}{A bootEGA object}

\item{vsize}{An integer indicating the size of the nodes. Default vsize = 6.}

\item{...}{Arguments passed to 'qgraph'.}
}
\description{
\code{plot.bootEGA} Plots the bootEGA typical structure using \code{\link{qgraph}}
}
\examples{
\dontrun{
ega.wmt <- EGA(data = wmt2[,7:24], model = "glasso")
boot.wmt <- bootEGA(data = wmt2[,7:24], n = 10, typicalStructure = TRUE,
plot.typicalStructure = TRUE, model = "GGM",
type = "parametric", ncores = 4, confirm = ega.wmt$wc)
plot.bootEGA(boot.wmt)
}

\dontrun{
plot(bootEGA)
}
}
\seealso{
\code{\link{EGA}} to estimate the number of dimensions of an instrument using EGA and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
