% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.EGA.R
\name{plot.EGA}
\alias{plot.EGA}
\title{Plot method for EGA objects.}
\usage{
\method{plot}{EGA}(ega.obj, title = "", vsize = 30, opacity = 0.4)
}
\arguments{
\item{ega.obj}{An EGA object}

\item{title}{Character. Title of the plot}

\item{vsize}{An integer indicating the size of the nodes. Default vsize = 30.}

\item{opacity}{A numeric value indicating the opacity of the edges. Default opacity = 0.4}
}
\description{
\code{plot} Plots the EGA result using \code{\link{qgraph}}
}
\examples{
ega.wmt <- EGA(data = wmt2[,7:24], plot.EGA = TRUE)
summary(ega.wmt)
plot(ega.wmt, title = "", vsize = 30, opacity = 0.4)

\dontrun{
plot(EGA)
}
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
\seealso{
\code{\link{EGA}} to estimate the number of dimensions of an instrument using EGA and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}

