% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dynEGA.Individuals.R
\name{print.dynEGA.Individuals}
\alias{print.dynEGA.Individuals}
\title{Print method for \code{\link[EGAnet]{dynEGA}} objects (Fixed Effects - Intraindividual Structure)}
\usage{
\method{print}{dynEGA.Individuals}(x, ...)
}
\arguments{
\item{x}{An \code{\link[EGAnet]{dynEGA}} objects (Fixed Effects - Intraindividual Structure)}

\item{...}{potentially further arguments (\strong{unused currently})}
}
\description{
Returns a summary of the \code{\link[EGAnet]{dynEGA}} objects (Fixed Effects - Intraindividual Structure)
}
\examples{
\dontrun{
# Estimate dynEGA
dyn.individual <- dynEGA(data = sim.dynEGA, n.embed = 5, tau = 1,
delta = 1, id = 21, group = 22, use.derivatives = 1,
effects = "fixed", model = "glasso")

#Print dynEGA results
print(dyn.individual)
}

}
\seealso{
\code{\link[EGAnet]{dynEGA}} to estimate the number of dimensions in multivariate time series using dynEGA.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
