% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.EGA.R
\name{print.EGA}
\alias{print.EGA}
\title{Print method for EGA objects.}
\usage{
\method{print}{EGA}(object)
}
\arguments{
\item{object}{An EGA object}
}
\description{
\code{print} Returns a summary of the EGA results
}
\examples{
ega.wmt <- EGA(data = wmt2[,7:24], plot.EGA = TRUE)
print(ega.wmt)

\dontrun{
print(EGA)
}
}
\seealso{
\code{\link{EGA}} to estimate the number of dimensions of an instrument using EGA and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
