% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{prints}
\alias{prints}
\alias{print.dynEGA}
\alias{print.dynEGA.Groups}
\alias{print.dynEGA.Individuals}
\alias{print.EGA}
\alias{print.NetLoads}
\title{S3Methods for Printing}
\usage{
\method{print}{dynEGA}(x,  ...)

\method{print}{dynEGA.Groups}(x, ...)

\method{print}{dynEGA.Individuals}(x, ...)

\method{print}{EGA}(x,  ...)

\method{print}{NetLoads}(x, ...)
}
\arguments{
\item{x}{Object from \code{EGAnet} package}

\item{...}{Additional arguments}
}
\value{
Prints \code{EGAnet} object
}
\description{
Prints for \code{EGAnet} objects
}
\author{
Hudson Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
