% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riEGA.R
\name{riEGA}
\alias{riEGA}
\title{Random-Intercept \code{\link[EGAnet]{EGA}}}
\usage{
riEGA(
  data,
  n = NULL,
  corr = c("auto", "pearson", "spearman"),
  na.data = c("pairwise", "listwise"),
  model = c("glasso", "TMFG"),
  algorithm = c("leiden", "louvain", "walktrap"),
  uni.method = c("expand", "LE", "louvain"),
  estimator = c("auto", "WLSMV", "MLR"),
  plot.EGA = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Numeric matrix or data frame.
Must be data representing \emph{only} the \strong{unrecoded} variables of interest.
\strong{Cannot} be a correlation matrix. Data that are not numeric will be
removed from the dataset}

\item{corr}{Character (length = 1).
Method to compute correlations.
Defaults to \code{"auto"} to automatically compute
appropriate correlations using \code{\link[EGAnet]{auto.correlate}}.
\code{"pearson"} and \code{"spearman"} are provide for completeness.
For other similarity measures, compute them first and input them
into \code{data} with the sample size (\code{n})}

\item{na.data}{Character (length = 1).
How should missing data be handled?
Available options:

\itemize{

\item{\code{"pairwise"}}
{Computes correlation for all available cases between
two variables}

\item{\code{"listwise"}}
{Computes correlation for all complete cases in the dataset}

}}

\item{model}{Character (length = 1).
Available options:

\itemize{

\item{\code{"BGGM"}}
{Computes the Bayesian Gaussian Graphical Model.
Set argument \code{ordinal.categories} to determine
levels allowed for a variable to be considered ordinal.
See \code{\link[BGGM]{estimate}} for more details}

\item{\code{"glasso"}}
{Computes the GLASSO with EBIC model selection.
See \code{\link[EGAnet]{EBICglasso.qgraph}} for more details}

\item{\code{"TMFG"}}
{Computes the TMFG method.
See \code{\link[EGAnet]{TMFG}} for more details}

}}

\item{algorithm}{Character or \code{\link{igraph}} \code{cluster_*} function.
Three options are listed below but all are available
(see \code{\link[EGAnet]{community.detection}} for other options):

\itemize{

\item{\code{"leiden"}}
{See \code{\link[igraph]{cluster_leiden}} for more details}

\item{\code{"louvain"}}
{See \code{\link[EGAnet]{community.consensus}} for more details.
By default, searches along resolutions from 0 to 2 in 0.05 increments
(\code{resolution_parameter = seq.int(0, 2, 0.05)}). Use the argument \code{resolution_parameter}
to change the search parameters (see examples)}

\item{\code{"walktrap"}}
{This algorithm is the default. See \code{\link[EGAnet]{cluster_walktrap}} for more details}

}}

\item{uni.method}{Character.
What unidimensionality method should be used? 
Defaults to \code{"louvain"}.
Current options are:

\itemize{

\item{\strong{\code{expand}}}
{Expands the correlation matrix with four variables correlated .50.
If number of dimension returns 2 or less in check, then the data 
are unidimensional; otherwise, regular EGA with no matrix
expansion is used. This is the method used in the Golino et al. (2020)
\emph{Psychological Methods} simulation.}

\item{\strong{\code{LE}}}
{Applies the Leading Eigenvalue algorithm (\code{\link[igraph]{cluster_leading_eigen}})
on the empirical correlation matrix. If the number of dimensions is 1,
then the Leading Eigenvalue solution is used; otherwise, regular EGA
is used. This is the final method used in the Christensen, Garrido,
and Golino (2021) simulation.}

\item{\strong{\code{louvain}}}
{Applies the Louvain algorithm (\code{\link[igraph]{cluster_louvain}})
on the empirical correlation matrix using a resolution parameter = 0.95.
If the number of dimensions is 1, then the Louvain solution is used; otherwise,
regular EGA is used. This method was validated in the Christensen (2022) simulation.}

}}

\item{estimator}{Character.
Estimator to use for random-intercept model (see \href{https://lavaan.ugent.be/tutorial/est.html}{Estimators}
for more details).
Defaults to \code{"auto"}, which selects \code{"MLR"} for continuous data and
\code{"WLSMV"} for mixed and categorical data.
Data are considered continuous data if they have 8 or
more categories (see Rhemtulla, Brosseau-Liard, & Savalei, 2012)}

\item{plot.EGA}{Boolean.
If \code{TRUE}, returns a plot of the network and its estimated dimensions.
Defaults to \code{TRUE}}

\item{verbose}{Boolean.
Whether messages and (insignificant) warnings should be output.
Defaults to \code{FALSE} (silent calls).
Set to \code{TRUE} to see all messages and warnings for every function call}

\item{...}{Additional arguments to be passed on to
\code{\link[EGAnet]{auto.correlate}}, \code{\link[EGAnet]{network.estimation}},
\code{\link[EGAnet]{community.detection}}, and
\code{\link[EGAnet]{community.consensus}}}
}
\value{
Returns a list containing:

\item{EGA}{Results from \code{\link[EGAnet]{EGA}}}

\item{RI}{A list containing information about the random-intercept
model (if the model converged):

\itemize{

\item{fit}
{The fit object for the random-intercept model using \code{\link[lavaan]{cfa}}}

\item{lavaan.args}
{The arguments used in \code{\link[lavaan]{cfa}}}

\item{loadings}
{Standardized loadings from the random-intercept model}

\item{correlation}
{Residual correlations after accounting for the random-intercept model}

}

}
}
\description{
Estimates the number of substantive dimensions after controlling for wording effects. 
EGA is applied to a residual correlation matrix after subtracting and random intercept
factor with equal unstandardized loadings from all the regular and unrecoded
reversed items in the database
}
\examples{
# Obtain example data
wmt <- wmt2[,7:24]

# riEGA example
riEGA(data = wmt, plot.EGA = FALSE)
# no plot for CRAN checks

}
\references{
# Selection of CFA Estimator \cr
Rhemtulla, M., Brosseau-Liard, P. E., & Savalei, V. (2012).
When can categorical variables be treated as continuous? A comparison of robust continuous and categorical SEM estimation methods under suboptimal conditions.
\emph{Psychological Methods}, \emph{17}, 354-373.
}
\seealso{
\code{\link[EGAnet]{plot.EGAnet}} for plot usage in \code{\link{EGAnet}}
}
\author{
Alejandro Garcia-Pardina <alejandrogp97@gmail.com>,
Francisco J. Abad <fjose.abad@uam.es>,
Alexander P. Christensen <alexpaulchristensen@gmail.com>,
Hudson Golino <hfg9s at virginia.edu>,
Luis Eduardo Garrido <luisgarrido@pucmm.edu.do>, and
Robert Moulder <rgm4fd@virginia.edu>
}
