% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.dynEGA.R
\docType{data}
\name{sim.dynEGA}
\alias{sim.dynEGA}
\title{sim.dynEGA Data}
\format{
A 10000x22 multivariate time series
}
\usage{
data(sim.dynEGA)
}
\description{
A simulated (multivariate time series) data with 20 variables, 200 individual observations,
50 time points per individual and 2 groups of individuals.
}
\examples{
data("sim.dynEGA")

}
\keyword{datasets}
