% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarys.R
\name{summarys}
\alias{summarys}
\alias{summary.hierEGA}
\title{S3Methods for summarying}
\usage{
\method{summary}{hierEGA}(object, ...)
}
\arguments{
\item{object}{Object from \code{EGAnet} package}

\item{...}{Additional arguments}
}
\value{
summarys \code{EGAnet} object
}
\description{
summarys for \code{EGAnet} objects
}
\author{
Hudson Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
