% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarys.R
\name{summarys}
\alias{summarys}
\alias{summary.dynEGA}
\alias{summary.dynEGA.Groups}
\alias{summary.dynEGA.Individuals}
\alias{summary.EGA}
\alias{summary.NetLoads}
\alias{summary.invariance}
\alias{summary.hierEGA}
\alias{summary.riEGA}
\title{S3Methods for summarying}
\usage{
\method{summary}{dynEGA}(object,  ...)

\method{summary}{dynEGA.Groups}(object, ...)

\method{summary}{dynEGA.Individuals}(object, ...)

\method{summary}{EGA}(object,  ...)

\method{summary}{NetLoads}(object, ...)

\method{summary}{invariance}(object, ...)

\method{summary}{hierEGA}(object, ...)

\method{summary}{riEGA}(object, ...)
}
\arguments{
\item{object}{Object from \code{EGAnet} package}

\item{...}{Additional arguments}
}
\value{
summarys \code{EGAnet} object
}
\description{
summarys for \code{EGAnet} objects
}
\author{
Hudson Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
