% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalCorMat.R
\name{totalCorMat}
\alias{totalCorMat}
\title{Total Correlation Matrix}
\usage{
totalCorMat(data)
}
\arguments{
\item{data}{Matrix or data frame.
Variables to be used in the analysis}
}
\value{
Returns a square matrix with pairwise total correlations
}
\description{
Computes the pairwise total correlation for a dataset
}
\examples{
# Compute total correlation matrix
totalCorMat(wmt2[,7:24])

}
\references{
Watanabe, S. (1960).
Information theoretical analysis of multivariate correlation.
\emph{IBM Journal of Research and Development} \emph{4}, 66-82.

# Implementation
Felix, L. M., Mansur-Alves, M., Teles, M., Jamison, L., & Golino, H. (2021).
Longitudinal impact and effects of booster sessions in a cognitive training program for healthy older adults.
\emph{Archives of Gerontology and Geriatrics}, \emph{94}, 104337.
}
\author{
Hudson F. Golino <hfg9s at virginia.edu>
}
