% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.example.R
\docType{data}
\name{toy.example}
\alias{toy.example}
\title{Toy Example}
\format{A 500x6 response matrix}
\usage{
data(toy.example)
}
\description{
A simulated dataset with 2 factors, three items per factor and  n = 500.
}
\examples{
data("toy.example")

}
\keyword{datasets}
